refactor: rename util.ts to index.ts
