+++ src/utils/fetchBlocks.ts
@@ -2,7 +2,7 @@ import { ethers } from 'ethers'
 import { IBlock } from 'interfaces/block.interface'
 import Block from '../models/block.model'
 import { logger } from './logger'
-import { getDateFromUnixTimestamp } from './util'
+import { getDateFromUnixTimestamp } from '.'
 
 const providerURL = process.env.PROVIDER_URL || 'https://rpc.gaiaxtestnet.oceanprotocol.com' // or use your local node 'http://localhost:8545'
 const provider = new ethers.providers.JsonRpcProvider(providerURL)
